# swagger/swagger/swagger.yml
openapi: 3.0.1
info:
  title: My API
  description: |
    # Overview 

    API documentation for Spring Boot application
    ## Structure api response format

    ### Base Api

    When a request sent to server, it always return a response has body like:

    ```json
    {
      "code": 401,
      "message": "Login"
      "data": Object
    }
    ```
    - `error`: status code
    - `message`: description about api
    - `data`: data will return for client, data structure is general

    ### Validation Api
    When a request failed because some field in body is missing, empty or not valid,
    API will return a response with **status: 422 (Unprocessable entity)**

    ```json
    {
      "code": 422,
      "message": "Validation Error",
      "error": [
        {
          "field": "field target"
          "message": "message"
        },
       {
          "field": "field target"
          "message": "message"
        }
      ]
    }
    ```
    - `code`: status code
    - `message`: description about api
    - `error`: contain array present field and message error
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /user:
    get:
      summary: Get info user
      description: Get info user when after login, return info to view in user card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                access-token:
                  type: string
                  description: bearer user access token
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: "#/components/schemas/ResponseApi"
                  - type: object
                    properties:
                      data:
                        properties:
                          id:
                            required: true
                            type: string
                            description: user id
                          name:
                            required: true
                            type: string
                            description: user id
                          uniId:
                            required: true
                            type: string
                            description: university id
                          className:
                            required: false
                            type: string
                            description: class room id
                          avatar:
                            required: false
                    example:
                      code: 200
                      message: "get info user"
                      data:
                        id: "0978-999-111"
                        name: "Le Anh Duc"
                        uniId: "21130320"
                        className: "DH21DTX"
                        avatar: "https://res.cloudinary.com/yourstyle/image/upload/c_scale/q_auto/f_auto/v1/product_img/4/Screenshot%202024-03-24%20170336.png"
  # -----------------------------------------------
  /user/check:
    post:
      summary: Checking user with uniId has existed in workspace id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uniId:
                  type: string
                  description: university id
                workspaceId:
                  type: string
                  description: workspace id
      responses:
        200:
          description: user with uniId is not exist in workspace id
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: "#/components/schemas/ResponseApi"
                  - type: object
        404:
          description: user with uniId is exist in workspace id
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: "#/components/schemas/ResponseApi"
                  - type: object
  # -----------------------------------------------
...
  /issue:
    post:
      summary: create issue
      description: create issue by student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: issue name
                sprintId:
                  type: string
                  description: sprint id
                projectId:
                  type: string
                  description: project id
                workspaceId:
                  type: string
                  description: workspace id
                storyPoint:
                  type: integer
                  description: story point
                  required: true
                status:
                  allOf:
                    - $ref: "#/components/schemas/Status"
                  description: status
                priority:
                  allOf:
                    - $ref: "#/components/schemas/Priority"
                  description: priority
                description:
                  type: string
                  description: description
                position:
                  type: number
                  description: "position of issue,
                    this is calculate in front-end and sent to server,
                    position = (prevPosition + nextPosition) / 2 = (2000 + 1000) /2"
                assignerId:
                  type: string
                  description: uniId of assigner
                reviewId:
                  type: string
                  description: uniId of reviewer
                start:
                  type: string
                  description: date time start
                end:
                  type: string
                  description: date time end
            example:
              name: "Design database"
              sprintId: "sprint-001"
              projectId: "project-001"
              workspaceId: "workspace-001"
              status: "Backlog"
              priority: "Major"
              description: "Lorem"
              position: 1200
              assignerId: "uni-002"
              reviewerId: "uni-001"
              start: ""
              end: ""

      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: "#/components/schemas/ResponseApi"
                  - type: object
              example:
                code: 200
                message: "Create issue"
    put:
      summary: update issue
      description: update issue by student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: issue id
                  required: true
                name:
                  type: string
                  description: issue name
                  required: true
                description:
                  type: string
                  description: issue description
                  required: true
                status:
                  $ref: "#/components/schemas/Status"
                storyPoint:
                  type: integer
                  description: story point assigned to issue
                  required: true
                priority:
                  type: string
                  description: value allow in Critical, Major, Minor, Trivial, Blocked
                  required: false
                  default: Critical
                start:
                  type: number
                  description: date time start issue
                  required: true
                end:
                  type: number
                  description: date time end issue
                  required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: "#/components/schemas/ResponseApi"
                  - type: object

              example:
                code: 200
                message: "Update Issue"
                data:
                  id: "project-001"
                  name: "project-001"
                  description: "..."
                  storyPoint: 10
                  priority: Critical
    patch:
      summary: change status issue (move card)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                issueId:
                  type: string
                  description: issue id
                  required: true
                status:
                  $ref: "#/components/schemas/Status"
                position:
                  type: number
                  description: use to decide order in list issue same status, front-end wil calculate it
              example:
                issueId: '10'
                status: "Todo"
                position: 2000
      responses:
        200:
          description: Success
        400:
          description: Not found issue with id
components:
  schemas:
...
    ResponseApi:
      $ref: "./components/response-api.yml"
...
    Status:
      type: string
      description: status issue, it will be a column name that contain issue
      default: Backlog
      enum:
        - Backlog
        - Todo
        - In Process
        - Review
        - Done
    Priority:
      type: string
      description: priority of issue
      default: Critical
      enum:
        - Critical
        - Major
        - Minor
        - Trivial
        - Blocked
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT